split Configuration.Certificates
